// apps/web/src/lib/api/generated/model/policyCreateCollectionMode.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * BrokerOS API
 * OpenAPI spec version: 1.0.0
 */

export type PolicyCreateCollectionMode = typeof PolicyCreateCollectionMode[keyof typeof PolicyCreateCollectionMode];


export const PolicyCreateCollectionMode = {
  insurer_direct: 'insurer_direct',
  agent_collected: 'agent_collected',
} as const;